+++ src/physics/raycast.c
@@ -14,15 +14,15 @@ RayCollision sphereRayTest(GFC_Edge3D ray, PhysicsBody *body) {
 	RayCollision col = {0};
 	GFC_Vector3D m;
 	gfc_vector3d_sub(m, p, body->position);
-	float b = gfc_vector3d_dot_product(m, d);
-	float c = gfc_vector3d_dot_product(m, m) - body->shape.shape.sphere.radius * body->shape.shape.sphere.radius;
+	double b = gfc_vector3d_dot_product(m, d);
+	double c = gfc_vector3d_dot_product(m, m) - body->shape.shape.sphere.radius * body->shape.shape.sphere.radius;
 	// Exit if r’s origin outside s (c > 0) and r pointing away from s (b > 0)
 	if (c > 0.0f && b > 0.0f) return col;
-	float discr = b*b - c;
+	double discr = b*b - c;
 	// A negative discriminant corresponds to ray missing sphere
 	if (discr < 0.0f) return col;
 	// Ray now found to intersect sphere, compute smallest t value of intersection
-	float t = -b - sqrtf(discr);
+	double t = -b - sqrt(discr);
 	// If t is negative, ray started inside sphere so clamp t to zero
 	if (t < 0.0f) t = 0.0f;
 	GFC_Vector3D q;
@@ -44,8 +44,8 @@ typedef struct AABB {
 // return intersection distance tmin and point q of intersection
 RayCollision intersectRayAABB(GFC_Vector3D p, GFC_Vector3D d, AABB a) {
 	RayCollision col = {0};
-	float tmin = 0.0;
-	float tmax = FLT_MAX;
+	double tmin = 0.0;
+	double tmax = DBL_MAX;
 	
 	// For all three slabs
 	for (int i = 0; i < 3; i++) {
@@ -54,9 +54,9 @@ RayCollision intersectRayAABB(GFC_Vector3D p, GFC_Vector3D d, AABB a) {
 			if (vec3_idx(p, i) < vec3_idx(a.min, i) || vec3_idx(p, i) > vec3_idx(a.max, i)) return col;
 		} else {
 			// Compute intersection t value of ray with near and far plane of slab
-			float ood = 1.0f / vec3_idx(d, i);
-			float t1 = (vec3_idx(a.min, i) - vec3_idx(p, i)) * ood;
-			float t2 = (vec3_idx(a.max, i) - vec3_idx(p, i)) * ood;
+			double ood = 1.0f / vec3_idx(d, i);
+			double t1 = (vec3_idx(a.min, i) - vec3_idx(p, i)) * ood;
+			double t2 = (vec3_idx(a.max, i) - vec3_idx(p, i)) * ood;
 			// Make t1 be intersection with near plane, t2 with far plane
 			if (t1 > t2) SWAP(t1, t2);
 			// Compute the intersection of slab intersection intervals
